skill-twin-engine/requirements.txt
# Core dependencies
streamlit==1.47.0
openai==2.6.1
python-dotenv==1.1.1
sentence-transformers==5.1.1
flask==3.0.0
flask-cors==4.0.0

# Web scraping dependencies
beautifulsoup4==4.13.5
requests==2.32.5
selenium==4.35.0
webdriver-manager==4.0.2

# Data processing
pandas==2.3.1
numpy==2.3.4

# Additional utilities
pdfplumber==0.11.4